+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/remote/client/RemoteClientManager.java
@@ -94,7 +94,7 @@ public class RemoteClientManager implements Service, ClusterWatcher {
 
     public void start() {
         Optional.ofNullable(sslContext).ifPresent(DynamicSslContext::start);
-        Executors.newSingleThreadScheduledExecutor()
+        Executors.newSingleThreadScheduledExecutor(r -> new Thread(r, "RemoteClientManager"))
                  .scheduleWithFixedDelay(new RunnableWithExceptionProtection(this::refresh, t -> log.error(
                      "Scheduled refresh Remote Clients failure.", t)), 1, 10, TimeUnit.SECONDS);
     }